requirements.txt
customtkinter>=5.2.0
pillow>=10.0.0
requests>=2.31.0
typing-extensions>=4.8.0